+++ apps/ddaas/src/app/[locale]/(content)/legal/layout.tsx
@@ -34,9 +34,6 @@ export default async function Layout({
     searchToggle: {
       enabled: false,
     },
-    themeSwitch: {
-      mode: 'light-only',
-    },
   };
 
   return (
@@ -61,7 +58,6 @@ export default async function Layout({
           tree: legalSource.getPageTree(locale),
           sidebar: { enabled: false },
           searchToggle: { enabled: false },
-          themeProvider: false,
         }}
       >
         {children}